+++ src/main/java/com/zeylin/pricetracker/dto/AddPriceRequest.java
@@ -3,6 +3,7 @@ package com.zeylin.pricetracker.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import javax.validation.constraints.NotNull;
 import java.time.LocalDate;
 
 @Getter
@@ -11,7 +12,9 @@ public class AddPriceRequest {
 
     private Integer itemId;
     private String itemName;
+    @NotNull
     private Integer amount;
+    @NotNull
     private LocalDate date;
     private Integer locationId;
     private String locationName;