+++ src/main/java/com/zeylin/pricetracker/utils/CustomExceptionHandler.java
@@ -2,14 +2,21 @@ package com.zeylin.pricetracker.utils;
 
 import com.zeylin.pricetracker.exceptions.InvalidArgumentException;
 import com.zeylin.pricetracker.exceptions.NotFoundException;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.ObjectUtils;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
 
 @ControllerAdvice
@@ -43,4 +50,28 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
         return new ResponseEntity<>(body, HttpStatus.BAD_REQUEST);
     }
 
+    /**
+     * Custom error message for @Valid validation errors.
+     */
+    @Override
+    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {
+        Map<String, Object> body = new LinkedHashMap<>();
+        body.put("timestamp", LocalDateTime.now());
+        body.put("status", HttpStatus.BAD_REQUEST.value());
+        body.put("error", HttpStatus.BAD_REQUEST.getReasonPhrase());
+        body.put("message", "Validation failed");
+
+        List<String> details = new ArrayList<>();
+        List<FieldError> fieldErrors = ex.getBindingResult().getFieldErrors();
+        for (FieldError error : fieldErrors) {
+            String msg = "Error on field '" + error.getField() +
+                    "': rejected value [" + ObjectUtils.nullSafeToString(error.getRejectedValue()) +
+                    "]; default message: " + error.getDefaultMessage();
+            details.add(msg);
+        }
+
+        body.put("details", details);
+
+        return new ResponseEntity<>(body, HttpStatus.BAD_REQUEST);
+    }
 }